+++ juniper/CHANGELOG.md
@@ -12,7 +12,7 @@ All user visible changes to `juniper` crate will be documented in this file. Thi
 
 ### BC Breaks
 
-- Upgraded [`chrono-tz` crate] integration to [0.9 version](https://github.com/chronotope/chrono-tz/releases/tag/v0.9.0). ([#1252])
+- Upgraded [`chrono-tz` crate] integration to [0.10 version](https://github.com/chronotope/chrono-tz/releases/tag/v0.10.0). ([#1252], [#1284])
 - Bumped up [MSRV] to 1.75. ([#1272])
 - Corrected compliance with newer [graphql-scalars.dev] specs: ([#1275], [#1277])
     - Switched `LocalDateTime` scalars to `yyyy-MM-ddTHH:mm:ss` format in types:
@@ -34,16 +34,19 @@ All user visible changes to `juniper` crate will be documented in this file. Thi
 
 ### Added
 
-- [`jiff` crate] integration behind `jiff` [Cargo feature]: ([#1271], [#1270])
+- [`jiff` crate] integration behind `jiff` [Cargo feature]: ([#1271], [#1278], [#1270])
     - `jiff::civil::Date` as `LocalDate` scalar.
     - `jiff::civil::Time` as `LocalTime` scalar.
     - `jiff::civil::DateTime` as `LocalDateTime` scalar. ([#1275])
     - `jiff::Timestamp` as `DateTime` scalar.
+    - `jiff::Zoned` as `ZonedDateTime` scalar.
+    - `jiff::tz::TimeZone` as `TimeZoneOrUtcOffset` and `TimeZone` scalars.
+    - `jiff::tz::Offset` as `UtcOffset` scalar.
     - `jiff::Span` as `Duration` scalar.
 
 ### Changed
 
-- Updated [GraphiQL] to [3.7.0 version](https://github.com/graphql/graphiql/blob/graphiql%403.7.0/packages/graphiql/CHANGELOG.md#370). ([#1279])
+- Updated [GraphiQL] to [3.7.1 version](https://github.com/graphql/graphiql/blob/graphiql%403.7.1/packages/graphiql/CHANGELOG.md#371). ([#1279])
 
 [#1252]: /../../pull/1252
 [#1270]: /../../issues/1270
@@ -51,7 +54,10 @@ All user visible changes to `juniper` crate will be documented in this file. Thi
 [#1272]: /../../pull/1272
 [#1275]: /../../pull/1275
 [#1277]: /../../pull/1277
+[#1278]: /../../pull/1278
 [#1279]: /../../pull/1279
+[#1281]: /../../pull/1281
+[#1284]: /../../pull/1284